+++ src/app/latte-chat/user/onboarding/page.tsx
@@ -1,0 +1,5 @@
+import OnboardingContainer from '@/features/user-onboarding/containers/OnboardingContainer'
+
+export default function UserOnboardingPage() {
+  return <OnboardingContainer />
+}
