--- src/app/latte-chat/user/onboarding/role/page.tsx
@@ -1,7 +1,0 @@
-'use client'
-
-import UserOnBoardingRoleContainer from './UserOnBoardingRoleContainer'
-
-export default function UserOnboardingRolePage() {
-  return <UserOnBoardingRoleContainer />
-}
