+++ src/features/user-onboarding/components/Checkbox/index.tsx
