+++ src/features/user-onboarding/components/GuideItem/index.tsx
