+++ src/features/user-onboarding/hooks/useSaveJuniorUserMutation.ts
@@ -1,0 +1,9 @@
+import { useMutation } from '@tanstack/react-query'
+import { saveJuniorUser } from '../services/onboardingService.client'
+
+export default function useSaveJuniorUserMutation() {
+  return useMutation({
+    mutationFn: ({ memberId, body }: { memberId: number; body: any }) =>
+      saveJuniorUser({ memberId }, body),
+  })
+}
