+++ src/features/user-onboarding/hooks/useSaveSeniorUserMutation.ts
@@ -1,0 +1,9 @@
+import { useMutation } from '@tanstack/react-query'
+import { saveSeniorUser } from '../services/onboardingService.client'
+
+export default function useSaveSeniorUserMutation() {
+  return useMutation({
+    mutationFn: ({ memberId, body }: { memberId: number; body: any }) =>
+      saveSeniorUser({ memberId }, body),
+  })
+}
